+++ src/pages/favorites-empty-page.tsx
