+++ src/pages/favorites-page.tsx
