+++ src/pages/offer-page.tsx
